app/requirements.txt
Flask==3.0.3
openstacksdk
